% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differentiation.R
\name{fast_jac}
\alias{fast_jac}
\title{Fast Numerical Jacobian}
\usage{
fast_jac(f_res, x, diff_method = c("forward", "central"), ...)
}
\arguments{
\item{f_res}{Function. A function returning a vector of residuals.}

\item{x}{Numeric vector. Parameters at which to evaluate the Jacobian.}

\item{diff_method}{String. Differentiation method: "forward" or "central".}

\item{...}{Additional arguments passed to \code{f_res}.}
}
\value{
A Jacobian matrix of dimension (m x n).
}
\description{
Calculates the Jacobian matrix for a vector-valued function (e.g., residuals).
}
